chat/client-libraries/cloud/requirements.txt
google_auth_oauthlib==1.2.0
protobuf==4.21.12
google-apps-chat==0.1.0
